„Днес“ по часовника на компютъра, не по UTC
Между 00:00 и 03:00 българско време UTC е още вчера. Възпроизведено с
подменен часовник (libfaketime, 00:30 Europe/Sofia): today() и
date('now') дават 2026-09-24 вместо 2026-09-25 — заемане след полунощ
с вчерашна дата, падеж „вчера“ още не е просрочен, копието носи
вчерашна дата, напомнянето и анулираният акт — вчерашния ден и часа в
Гринуич.

- local-date.js: localDate(); main.js today(), дневник, копия, каталог,
  MARC 801, резервните today() в два обработчика
- src/views/core.js: today() местна; tsDay()/tsTime() за UTC моменти от
  базата (напомняния, анулиран/съставен акт, КДБФ)
- SQL: date('now') → date('now', 'localtime') в 11 файла; datetime('now')
  (моменти в UTC) остава; смятането с дати (addDays, седмици) — в UTC
- тестовете: общ test/helpers/local-day.js за еталоните „днес“, иначе
  нов прозорец 00:00–03:00 за плаващи падания в CI (Europe/Sofia)
- inventory_sessions.free_access_pct е INTEGER, както в settings

Как беше проверено: test/proverka-v2467.test.js — дете-процес в зона, в
която местната дата СЕГА е различна от UTC (UTC−12 или UTC+14 според
часа): заемане с местната дата и „Просрочени“ през истинския main.js;
екранът (today, tsDay, tsTime) в jsdom; пазач за голо date('now').
Осем мутации — всяка уловена. Цялата поредица при 00:30 софийско време:
2093/2096; трите падащи (копия) падат и със стария код, и в 12:00 при
подменен часовник — времената на файловете идват от истинския часовник.

+++ electron-app/db/schema.sql
@@ -490,7 +490,7 @@ CREATE TABLE IF NOT EXISTS inventory_sessions (
   -- пренаписваше норматива на вече подписан протокол — 7 липсващи при допустими 5
   -- („над норматива, прилага се редът по чл. 51 – 53“) ставаха 7 при допустими 10.
   -- NULL = неприключена сесия (тогава се чете текущата настройка).
-  free_access_pct REAL
+  free_access_pct INTEGER
 );
 CREATE TABLE IF NOT EXISTS inventory_session_scans (
   id          INTEGER PRIMARY KEY AUTOINCREMENT,
